feat: ProductList 재활용을 위한 props 분리
